Raw: Updates projects architecture
