+++ legacy/solucaopdfplumber.py
